+++ CHANGELOG.md
@@ -22,6 +22,7 @@ event which has previously failing was resolved
 - TCP/UDP sockets now accept 1.x backward compatible payloads. 1.x Check Result gets translated to a 2.x Event.
 - Custom attributes can be added to the agent at start.
 - New and improved Check Hooks are implemented (see whats new about hooks here: [Hooks](https://github.com/sensu/sensu-alpha-documentation/blob/master/08-hooks.md))
+- Add check subdue CLI support.
 
 ### Changed
 - Avoid using reflection in time.InWindows function.